+++ xarray/core/dataarray.py
@@ -2114,19 +2114,6 @@ class DataArray(AbstractArray, DataWithCoords):
         ds = self._to_temp_dataset().rank(dim, pct=pct, keep_attrs=keep_attrs)
         return self._from_temp_dataset(ds)
 
-    def isin(self, test_elements):
-        if LooseVersion(np.__version__) < LooseVersion('1.13.0'):
-            raise ImportError('isin requires numpy version 1.13.0 or later')
-        from .computation import apply_ufunc
-
-        return apply_ufunc(
-            np.isin,
-            self,
-            kwargs=dict(test_elements=test_elements),
-            dask='parallelized',
-            output_types=[np.bool_],
-        )
-
 
 # priority most be higher than Variable to properly work with binary ufuncs
 ops.inject_all_ops_and_reduce_methods(DataArray, priority=60)
